requirements.txt
wheel
requests
click
numpy
tifffile
pillow
python-dateutil
utm
shapely
boto3
botocore
pyproj
oauthlib
requests_oauthlib
attrs>=18.2.0
aenum>=2.1.4
